requirements.txt
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
backports.ssl-match-hostname==3.4.0.2
boto==2.38.0
Brlapi==0.6.1
bz2file==0.98
certifi==2015.4.28
chardet==2.0.1
cmd2==0.6.8
colorama==0.2.5
command-not-found==0.3
configglue==1.1.2
Cython==0.23.3
debtagshw==0.1
decorator==3.4.0
deepdish==0.3.2
defer==1.0.6
dirspec==13.10
docopt==0.6.2
duplicity==0.6.23
fake-factory==0.5.0
flake8==2.3.0
Flask==0.10.1
fuzzywuzzy==0.5.0
gdbn==0.1
gensim==0.12.3
gitdb==0.6.4
GitPython==1.0.0
gnumpy==0.2
GnuPGInterface==0.3.2
gnureadline==6.3.3
gunicorn==19.3.0
h5py==2.5.0
html5lib==0.999
httplib2==0.8
httpretty==0.8.6
iotop==0.6
ipython==3.1.0
itsdangerous==0.24
Jinja2==2.7.3
joblib==0.8.0a3
jsonschema==2.4.0
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lda==1.0.3
leveldb==0.193
line-profiler==1.0
lmdb==0.83
lockfile==0.8
louis==2.5.3
lxml==3.3.3
Magic-file-extensions==0.2
Mako==0.9.1
MarkupSafe==0.18
matplotlib==1.3.1
matplotlib2tikz==0.1.0
mccabe==0.3
mercurial==2.8.2
mistune==0.5.1
mpi4py==1.3.1
munkres==1.0.6
natsort==4.0.4
networkx==1.8.1
nibabel==2.0.1
nltk==3.1
nolearn==0.5
nose==1.3.1
numexpr==2.2.2
numpy==1.11.0
oauth==1.0.1
oauthlib==0.6.1
oneconf==0.3.7.14.4.1
PAM==0.4.2
pandas==0.13.1
pandoc==1.0.0a3
path.py==8.2.1
pbr==1.8.1
pep8==1.5.7
pexpect==3.1
Pillow==2.4.0
piston-mini-client==0.7.5
plyvel==0.9
protobuf==3.0.0b3
PsiTurk==2.1.2
psutil==2.2.1
ptyprocess==0.5
Pycluster==1.52
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pydot==1.0.28
pyemd==0.0.10
pyflakes==0.8.1
Pygments==2.0.1
pygobject==3.12.0
pyinotify==0.9.4
pyIPCA==1.0
pykalman==0.9.5
-e git://github.com/lisa-lab/pylearn2.git@f85582d20e5ec826dc603aeac2642981caac94e6#egg=pylearn2-master
pyOpenSSL==0.13
pyparsing==2.0.3
pyserial==2.6
pysmbc==1.0.14.1
-e git://git.code.sf.net/p/pysparse/git@1751931aa5c06429245c3aca3a3ba6d7622ea8b5#egg=pysparse-master
python-apt===0.9.3.5ubuntu1
python-dateutil==1.5
python-debian===0.1.21-nmu2ubuntu2
python-gflags==2.0
python-Levenshtein==0.12.0
pytz==2012rc0
pyxdg==0.25
PyYAML==3.11
pyzmq==14.1.1
reportlab==3.0
requests==2.6.2
rhythmbox-ubuntuone==4.2.0
scikit-image==0.9.3
scikit-learn==0.14.1
scipy==0.16.0
SecretStorage==2.0.0
sentiment-classifier==0.6
sessioninstaller==0.0.0
setproctitle==1.1.8
sh==1.11
simplejson==3.3.1
six==1.10.0
smart-open==1.3.0
smmap==0.9.0
software-center-aptd-plugins==0.0.0
spams==2.4
SQLAlchemy==1.0.2
ssh-import-id==3.21
stop-words==2015.2.23.1
system-service==0.1.6
tables==3.1.1
tensorflow==0.9.0rc0
terminado==0.5
textblob==0.11.0
textmining==1.0
Theano==0.7.0
tornado==4.2
trash-cli==0.12.7
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
ua-parser==0.3.6
ujson==1.35
unity-lens-photos==1.0
urllib3==1.7.1
user-agents==0.3.2
virtkey==0.63.0
virtualenv==13.1.0
wadllib==1.3.2
Werkzeug==0.10.4
wheel==0.24.0
xdiagnose===3.6.3build2
xkit==0.0.0
zope.interface==4.0.5
